requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
email-validator==2.1.0
google-generativeai==0.3.2
pinecone-client==3.0.0
langchain==0.1.0
langchain-google-genai==0.0.8
numpy==1.24.3
tiktoken==0.5.2